Rename SetWorld to WorldSetter
