+++ src/lib.rs
@@ -450,7 +450,7 @@ impl PluginGroup for DefaultTweenPlugins {
             .add(TweenCorePlugin::default())
             .add(register_types);
 
-        let group = group.add(set::SetWorldPlugin);
+        let group = group.add(set::WorldSetterPlugin);
 
         let group = group
             .add(set::component::<items::Translation>())
